Move tokeniser to tokeniser directory, change package name
