+++ tokeniser/tokeniser_test.go
@@ -1,4 +1,4 @@
-package rapid
+package tokeniser
 
 import (
 	"testing"
